Fix with casting
Realized I needed to cast the long to a double so I could get the
appropriate fraction in the double.

+++ FT2016_Sprint2/src/org/usfirst/frc/team3459/robot/PT_Timer.java
@@ -32,7 +32,7 @@ public class PT_Timer {
    * Get the number of seconds since start time
    */
   public double getSecs(){
-	  return (getMillis() / 1000);
+	  return (((double)getMillis()) / 1000);
   }
     
 }
